Improve handling of same-day updates.
The check for the current day's episode had a flaw in that if you restarted the
tool after an update added the current day, but the show had not yet started,
it would do short polling for the episode.

Instead of checking for the start of the episode, check that the latest has the
same date as the current moment, as the indicator for long polling.