epdate: Retry on missing video ID.
It appears sometimes there is a lag between when the video ID is created and
when it is available for query. Occasionally the announcement tweet may go out
in that window, and the script doesn't fetch the description properly.  To work
around this, treat a missing ID as a retryable event.