epdate: Add -check-video flag.
Most of the time, we want to fail if the episode tweet does not contain a
YouTube video ID for us to link to, since that is where we get the description
of the episode.

Sometimes, however, the reason the video is missing is that the human who wrote
the tweet hamfingered the URL, and it's there but not recognized by the parser.
Then we want to still generate an episode file, and the operator can paste in
the link manually from the output.

This is greasy but entirely in keeping with the way all of this works.